requirements.txt
asgiref==3.3.1
cleanup==1.0.3
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-ordered-model==3.5
docopt==0.6.2
gunicorn==20.1.0
huepy==1.2.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
